% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.getCountQuants.dr.R
\name{boot.getCountQuants.dr}
\alias{boot.getCountQuants.dr}
\title{Obtain the bootstrap version of quantiles}
\usage{
boot.getCountQuants.dr(Y, X, t, y0, cts, B, quants = NULL)
}
\arguments{
\item{Y}{The outcome variable}

\item{X}{The control variables}

\item{t}{The treatment variable}

\item{y0}{The cutoff values}

\item{cts}{A list of counterfactual treatments}

\item{B}{The number of bootstrap repetions}

\item{quants}{The quantiles to be obtained}
}
\value{
The bootstrap version of quantiles
}
\description{
Obtain the bootstrap version of quantiles
}
