% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.getCountQuants.qr.R
\name{boot.getCountQuants.qr}
\alias{boot.getCountQuants.qr}
\title{Obtain the bootstrap version of quantiles}
\usage{
boot.getCountQuants.qr(B = NULL, Y, X, t, cts, quants = NULL, s = NULL,
  e = NULL)
}
\arguments{
\item{B}{The number of bootstrap repetions}

\item{Y}{The outcome variable}

\item{X}{The control variables}

\item{t}{The treatment variable}

\item{cts}{A list of counterfactual treatments}

\item{quants}{The quantiles to be obtain}

\item{s}{The number of linear quantile regressions}

\item{e}{The trimming level to avoid estimation of tail quantiles}
}
\value{
The bootstrap version of quantiles
}
\description{
Obtain the bootstrap version of quantiles
}
