% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{CFA}
\alias{CFA}
\title{title}
\usage{
CFA(tvals, distcondt, bootiterlist = NULL, tvallist = NULL, coef = NULL)
}
\arguments{
\item{tvals}{the values of the continuous treatment with which
to compute counterfactuals}

\item{distcondt}{an ecdf object for a particular value of the treatment}

\item{bootiterlist}{a list of bootstrapped CFA objects that can be used
for computing standard errors}

\item{tvallist}{the values of the treatment used in each bootstrap iteration}
}
\value{
CFA object
}
\description{
CFA objects
}
