% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R, R/imports.R
\docType{package}
\name{cfa}
\alias{cfa}
\alias{cfa}
\alias{cfa-package}
\title{cfa}
\usage{
cfa(tvals, yvals, data, yname, tname, xnames = NULL, drobj = NULL,
  se = TRUE, iters = 100, cl = 1)
}
\arguments{
\item{tvals}{the values of the continuous treatment with which
to compute counterfactuals}

\item{yvals}{the values to compute the counterfactual distribution fo}

\item{yname}{the name of the outcome (y) variable}

\item{tname}{the name of the treatment (t) variable}

\item{xnames}{the names of additional control variables to include}

\item{drobj}{optional distribution regression object that has been previously
computed}

\item{se}{whether or not to compute standard errors using the bootstrap}

\item{iters}{how many bootstrap iterations to use}

\item{cl}{how many clusters to use for parallel computation of standard
errors}
}
\value{
CFA object
}
\description{
compute counterfactuals using distribution regression
 with a continuous treatment

cfa: A package for computing counterfactuals with a continuous treatment
}
