% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{cfa2}
\alias{cfa2}
\title{cfa2}
\usage{
cfa2(tvals, yvals, data, yname, tname, xnames1 = NULL, drobj1 = NULL,
  xnames2 = NULL, drobj2 = NULL, se = TRUE, iters = 100, cl = 1)
}
\arguments{
\item{tvals}{the values of the continuous treatment with which
to compute counterfactuals}

\item{yvals}{the values to compute the counterfactual distribution fo}

\item{yname}{the name of the outcome (y) variable}

\item{tname}{the name of the treatment (t) variable}

\item{xnames1}{the first set of x variables}

\item{drobj1}{the first distribution regression object, optional}

\item{xnames2}{the second set of x variables}

\item{drobj2}{the second distribution regression object, optional}

\item{se}{whether or not to compute standard errors using the bootstrap}

\item{iters}{how many bootstrap iterations to use}

\item{cl}{how many clusters to use for parallel computation of standard
errors}
}
\value{
list of two CFA objects
}
\description{
the same as cfa method except it computes two results at the
 same time which allows one to conduct inference on their difference
}
