% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{CFASE}
\alias{CFASE}
\title{CFASE}
\usage{
CFASE(tvals, est, se = NULL, c = NULL)
}
\arguments{
\item{tvals}{the values of the "treatment" to compute parameters of
interest for}

\item{est}{the estimate of the parameter at each value of t}

\item{se}{whether or not to compute standard errors using the bootstrap}

\item{c}{optional critical value for uniform inference}
}
\value{
CFASE object
}
\description{
creates object of class CFASE
}
