% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Co.treat.plot.dr.R
\name{Co.treat.plot.dr}
\alias{Co.treat.plot.dr}
\title{Plot coefficients on treatment variable}
\usage{
Co.treat.plot.dr(CondDiFs, y0)
}
\arguments{
\item{CondDiFs}{The conditional distributions functions}

\item{y0}{The cutoff values}
}
\description{
Plot coefficients on treatment variable
}
