% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{compute.cfa}
\alias{compute.cfa}
\title{compute.cfa}
\usage{
compute.cfa(tvals, yvals, data, yname, tname, xnames = NULL, drobj = NULL)
}
\arguments{
\item{tvals}{the values of the continuous treatment with which
to compute counterfactuals}

\item{yvals}{the values to compute the counterfactual distribution fo}

\item{yname}{the name of the outcome (y) variable}

\item{tname}{the name of the treatment (t) variable}

\item{xnames}{the names of additional control variables to include}

\item{drobj}{optional distribution regression object that has been previously
computed}
}
\value{
CFA object
}
\description{
does the heavy lifting for computing cfa results
}
\keyword{internal}
