% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{compute.cfa2}
\alias{compute.cfa2}
\title{compute.cfa2}
\usage{
compute.cfa2(tvals, yvals, data, yname, tname, xnames = NULL, method = "dr",
  link = "logit", tau = seq(0.1, 0.99, 0.01), condDistobj = NULL)
}
\arguments{
\item{condDist}{optional pre-estimated conditional distribution function}
}
\value{
CFA.OBJ
}
\description{
an update of compute.cfa that allows one to use other
 other estimators such as quantile regression or nonparametric methods
 (not yet implemented) as the first step estimator of the conditional
 distribution
}
\keyword{internal}
