% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountDisFs.dr.plot.R
\name{CountDisFs.dr.plot}
\alias{CountDisFs.dr.plot}
\title{Plot the counterfactual distribution functions before rearrangement}
\usage{
CountDisFs.dr.plot(y0, CountDisFs.dr)
}
\arguments{
\item{y0}{The cutoff values}

\item{CountDisFs.dr}{The counterfactual distribution functions before rearrangement}
}
\description{
Plot the counterfactual distribution functions before rearrangement
}
