% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountQuants.dr.plot.R
\name{CountQuants.dr.plot}
\alias{CountQuants.dr.plot}
\title{Plot the counterfactual quantiles}
\usage{
CountQuants.dr.plot(cts, CountQuants.dr, quans)
}
\arguments{
\item{cts}{A list of counterfactual treatments. 'cts' must be chosen from CountQuants.dr matrix}

\item{CountQuants.dr}{The counterfactual quantiles}

\item{quans}{The quantiles to plot, 'quans' must be chosen from CountQuants.dr matrix}
}
\description{
Plot the counterfactual quantiles
}
