% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountQuants.plot.R
\name{CountQuants.plot}
\alias{CountQuants.plot}
\title{Plot the counterfactual quantiles}
\usage{
CountQuants.plot(CountQuants, SDMat)
}
\arguments{
\item{CountQuants}{The counterfactual quantiles matrix}

\item{SDMat}{The standard errors matrix}
}
\description{
Plot the counterfactual quantiles
}
