% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisFs2Quants.qr.R
\name{DisFs2Quants.qr}
\alias{DisFs2Quants.qr}
\title{Convert counterfactual distribution functions to quantiles}
\usage{
DisFs2Quants.qr(Y, X, t, ct, quants = NULL, s = NULL, e = NULL)
}
\arguments{
\item{Y}{The outcome variable}

\item{X}{The control variables}

\item{t}{The treatment variable}

\item{ct}{A counterfactual treatment level}

\item{quants}{The quantiles to be obtained}

\item{s}{The number of linear quantile regressions}

\item{e}{The trimming level to avoid estimation of tail quantiles}
}
\value{
The counterfactual quantiles for a treatment
}
\description{
Calculate the counterfactual quantiles for a treatment
}
\details{
This function
}
