% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{getCoef.CFA}
\alias{getCoef.CFA}
\title{getCoef.CFA}
\usage{
getCoef.CFA(cfaobj, yvals, se = T, ...)
}
\arguments{
\item{cfaobj}{a CFA object}

\item{se}{whether or not to compute standard errors}

\item{...}{can pass additional arguments to fun using this argument}

\item{fun}{a function to apply for every value of the treatment in the
cfaobj}

\item{setype}{whether to comput "pointwise" or "uniform" standard errors}
}
\value{
CFASE object
}
\description{
get a particular parameter of interest from a cfa object
}
