% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCondDiFs.dr.R
\name{getCondDiFs.dr}
\alias{getCondDiFs.dr}
\title{Obtain the conditional distribution functions}
\usage{
getCondDiFs.dr(Y, X, t, y0)
}
\arguments{
\item{Y}{The outcome variable}

\item{X}{The control variables}

\item{t}{The treatment variable}

\item{y0}{A vector of cutoff values}
}
\value{
The conditional distribution functions
}
\description{
Obtain the conditional distribution functions
}
