% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCondQuants.qr.R
\name{getCondQuants.qr}
\alias{getCondQuants.qr}
\title{Get conditional quantiles for a treatment}
\usage{
getCondQuants.qr(Y, X, t, s = NULL, e = NULL)
}
\arguments{
\item{Y}{The outcome variable}

\item{X}{The control variables}

\item{t}{The treatment variable}

\item{s}{The number of linear quantile regressions}

\item{e}{The trimming level to avoid estimation of tail quantiles}
}
\value{
The conditional quantiles
}
\description{
Get conditional quantiles for a treatment
}
