% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountQuant.qr.R
\name{getCountQuant.qr}
\alias{getCountQuant.qr}
\title{Obtain the counterfacutal quantiles for one treatment}
\usage{
getCountQuant.qr(Y, X, t, ct, s = NULL, e = NULL)
}
\arguments{
\item{Y}{The outcome variable}

\item{X}{The control variables}

\item{t}{The treatment variable}

\item{ct}{A counterfactual treatment level}

\item{s}{The number of linear quantile regressions}

\item{e}{The trimming level to avoid estimation of tail quantiles}
}
\value{
The counterfacutal quantiles for one treatment
}
\description{
Calculate the counterfactual quantiles
}
