% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountQuants.dr.R
\name{getCountQuants.dr}
\alias{getCountQuants.dr}
\title{Obtain the counterfactual quantiles}
\usage{
getCountQuants.dr(CondDiFs, y0, X, cts, quants = NULL)
}
\arguments{
\item{CondDiFs}{The conditional distributions functions}

\item{y0}{The cutoff values}

\item{X}{The control variables}

\item{cts}{A list of counterfactual treatments}

\item{quants}{The quantiles to be obtained}
}
\value{
The counterfactual quantiles
}
\description{
Obtain the counterfactual quantiles
}
