% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountQuants.qr.R
\name{getCountQuants.qr}
\alias{getCountQuants.qr}
\title{Get conditional quantiles for a list of treatments}
\usage{
getCountQuants.qr(Y, X, t, cts, quants = NULL, s = NULL, e = NULL)
}
\arguments{
\item{Y}{The outcome variable}

\item{X}{The control variables}

\item{t}{The treatment variable}

\item{cts}{A list of counterfactual treatments}

\item{quants}{The quantiles to be obtain}

\item{s}{The number of linear quantile regressions}

\item{e}{The trimming level to avoid estimation of tail quantiles}
}
\value{
The conditional quantiles for a list of treatments
}
\description{
Get conditional quantiles for a list of treatments
}
