% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{getRes.CFA}
\alias{getRes.CFA}
\title{getRes.CFA}
\usage{
getRes.CFA(cfaobj, fun, se = T, ...)
}
\arguments{
\item{cfaobj}{a CFA object}

\item{fun}{a function to apply for every value of the treatment in the
cfaobj}

\item{se}{whether or not to compute standard errors}

\item{...}{can pass additional arguments to fun using this argument}
}
\value{
CFASE object
}
\description{
get a particular parameter of interest from a cfa object
}
