% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{getResDiff.CFA}
\alias{getResDiff.CFA}
\title{getResDiff.CFA}
\usage{
getResDiff.CFA(cfaobj1, cfaobj2, fun, se = T, ...)
}
\arguments{
\item{cfaobj1}{the first CFA object}

\item{cfaobj2}{the second CFA object}

\item{fun}{a function to apply for every value of the treatment in the
cfaobj}

\item{se}{whether or not to compute standard errors}

\item{...}{can pass additional arguments to fun using this argument}
}
\value{
CFASE object
}
\description{
Get the difference between two CFA objects
}
