% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSDMat.dr.R
\name{getSDMat.dr}
\alias{getSDMat.dr}
\title{Obtain the standard errors matrix for quantiles by bootstrap}
\usage{
getSDMat.dr(Y, X, t, y0, cts, B, quants = NULL)
}
\arguments{
\item{Y}{The outcome variable}

\item{X}{The control variables}

\item{t}{The treatment variable}

\item{y0}{The cutoff values}

\item{cts}{A list of counterfactual treatments}

\item{B}{The number of bootstrap repetions}

\item{quants}{The quantiles to be obtained}
}
\value{
The standard errors matrix for quantiles
}
\description{
Obtain the standard errors matrix for quantiles by bootstrap
}
