% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSDMat.qr.R
\name{getSDMat.qr}
\alias{getSDMat.qr}
\title{Obtain the standard errors matrix for quantiles by bootstrap using quantile regression approach}
\usage{
getSDMat.qr(B = NULL, Y, X, t, cts, quants = NULL, s = NULL, e = NULL)
}
\arguments{
\item{B}{The number of bootstrap repetitions}

\item{Y}{The outcome variable}

\item{X}{The control variables}

\item{t}{The treatment variable}

\item{cts}{A list of counterfactual treatments}

\item{quants}{The quantiles to be obtain}

\item{s}{The number of linear quantile regressions}

\item{e}{The trimming level to avoid estimation of tail quantiles}
}
\value{
The standard errors matrix for quantiles using quantile regression approach
}
\description{
Obtain the standard errors matrix for quantiles by bootstrap using quantile regression approach
}
