% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{IQR}
\alias{IQR}
\title{IQR}
\usage{
IQR(edf, t1, t2)
}
\arguments{
\item{edf}{an ecdf}

\item{t1}{upper quantile}

\item{t2}{lower quantile}
}
\value{
scalar interquantile range
}
\description{
compute interquantile range from ecdf object
}
