% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alige.R
\name{localIGE.inner}
\alias{localIGE.inner}
\title{localIGE.inner}
\usage{
localIGE.inner(t, Y, T, Xmat, h)
}
\arguments{
\item{t}{conditional at a value T=t}

\item{Y}{outcome variable}

\item{T}{treatment variable}

\item{Xmat}{covariates}

\item{h}{bandwidth}
}
\value{
lige
}
\description{
local intergenerational elasticities
}
