% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alige.R
\name{localIGE}
\alias{localIGE}
\title{localIGE}
\usage{
localIGE(tvals, Y, T, Xmat, h, cl = 1)
}
\arguments{
\item{tvals}{a grid of values of treatment variable}

\item{Y}{outcome variable}

\item{T}{treatment variable}

\item{Xmat}{covariates}

\item{h}{bandwidth}

\item{cl}{number of cores to use}
}
\description{
localIGE
}
