% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCountDisFs.dr.R
\name{makeCountDisFs.dr}
\alias{makeCountDisFs.dr}
\title{Obtain the counterfactual distribution functions before rearrangement}
\usage{
makeCountDisFs.dr(CondDiFs, y0, X, cts)
}
\arguments{
\item{CondDiFs}{The conditional distributions functions}

\item{y0}{The cutoff values}

\item{X}{The control variables}

\item{cts}{A list of counterfactual treatments}
}
\value{
The counterfactual distribution functions before rearrangement
}
\description{
Obtain the counterfactual distribution functions before rearrangement
}
