% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCountDisFs.qr.R
\name{makeCountDisFs.qr}
\alias{makeCountDisFs.qr}
\title{Make counterfactual distribution functions}
\usage{
makeCountDisFs.qr(Y, X, t, ct, s = NULL, e = NULL)
}
\arguments{
\item{Y}{The outcome variable}

\item{X}{The control variables}

\item{t}{The treatment variable}

\item{ct}{A counterfactual treatment level}

\item{s}{The number of linear quantile regressions}

\item{e}{The trimming level to avoid estimation of tail quantiles}
}
\value{
The counterfactual distribution functions
}
\description{
Make counterfactual distribution functions
}
