% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{pov}
\alias{pov}
\title{pov}
\usage{
pov(edf, povline)
}
\arguments{
\item{edf}{an ecdf}

\item{povline}{the poverty line}
}
\value{
scalar fraction below poverty line
}
\description{
compute fraction below the poverty line from ecdf
}
