% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_CountDisFs.dr.plot.R
\name{re_CountDisFs.dr.plot}
\alias{re_CountDisFs.dr.plot}
\title{Plot the counterfactual distribution functions after rearrangement}
\usage{
re_CountDisFs.dr.plot(y0, re_CountDisFs.dr)
}
\arguments{
\item{y0}{The cutoff values}

\item{re_CountDisFs.dr}{The counterfactual distribution functions after rearrangement}
}
\description{
Plot the counterfactual distribution functions after rearrangement
}
