% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_makeCountDisFs.dr.R
\name{re_makeCountDisFs.dr}
\alias{re_makeCountDisFs.dr}
\title{Obtain the rearranged counterfactual distribution functions}
\usage{
re_makeCountDisFs.dr(CondDiFs, y0, X, cts)
}
\arguments{
\item{CondDiFs}{The conditional distributions functions}

\item{y0}{The cutoff values}

\item{X}{The control variables}

\item{cts}{A list of counterfactual treatments}
}
\value{
The counterfactual distribution functions after rearrangement
}
\description{
Obtain the rearranged counterfactual distribution functions
}
