% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{rich}
\alias{rich}
\title{rich}
\usage{
rich(edf, richline)
}
\arguments{
\item{edf}{an ecdf}

\item{richline}{the cutoff for being rich}
}
\value{
scalar fraction that are "rich"
}
\description{
compute fraction of "rich"
}
