% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alige.R
\name{summary_LLIGE}
\alias{summary_LLIGE}
\title{summary_LLIGE}
\usage{
summary_LLIGE(object, ...)
}
\arguments{
\item{object}{an \code{LLIGE} object}

\item{...}{extra arguments}
}
\value{
summary of \code{LLIGE}
}
\description{
prints a summary of a \code{LLIGE} object
}
\examples{
data(igm)
igm$hs=ifelse(igm$HEDUC=="HS",1,0)
igm$col=ifelse(igm$HEDUC=="COL",1,0)
formla=lcfincome~lfincome
xformla=~hs+col
tvals=seq(quantile(igm$lfincome,probs = 0.1),quantile(igm$lfincome,probs = 0.9),length.out = 10)
h=1.2
data=igm
B=7
cl=1
object=LLIGE(B,formla=formla, xformla=xformla, data=data,tvals=tvals,h=h,cl=cl)
summary_LLIGE(object)
}
