% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{test.CFA}
\alias{test.CFA}
\title{test.CFA}
\usage{
test.CFA(cfaobj, fun, allt, se = T, ...)
}
\arguments{
\item{cfaobj}{a CFA object}

\item{fun}{which function to use}

\item{allt}{all values of t in the dataset}

\item{se}{whether or not to compute standard errors}

\item{...}{additional parameters for the function fun}
}
\value{
CFASE object
}
\description{
test if a counterfactual distribution is equal to its average
 for all values of the treatment
}
