% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{getCoef.CFA}
\alias{getCoef.CFA}
\title{getCoef.CFA}
\usage{
getCoef.CFA(cfaobj, yvals, se = T, ...)
}
\arguments{
\item{cfaobj}{a CFA object}

\item{yvals}{the y values that the cfa object is computed for}

\item{se}{whether or not to compute standard errors}

\item{...}{can pass additional arguments to fun using this argument}
}
\value{
CFASE object
}
\description{
get a particular parameter of interest from a cfa object
}
