% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa.R
\name{getResDiff.CFA}
\alias{getResDiff.CFA}
\title{getResDiff.CFA}
\usage{
getResDiff.CFA(cfaobj1, cfaobj2, fun, se = T, ...)
}
\arguments{
\item{fun}{a function to apply for every value of the treatment in the
cfaobj}

\item{se}{whether or not to compute standard errors}

\item{...}{can pass additional arguments to fun using this argument}

\item{cfaobj}{a CFA object}

\item{setype}{whether to compute "pointwise" or "uniform" standard errors}
}
\value{
CFASE object
}
\description{
get a particular parameter of interest from a cfa object
}
