% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alige.R
\name{localIGE.inner}
\alias{localIGE.inner}
\title{localIGE.inner}
\usage{
localIGE.inner(formla, xformla = NULL, data, t, h = NULL)
}
\arguments{
\item{formla}{a formula y ~ treatment}

\item{xformla}{one sided formula for x variables to include, e.g. ~x1 + x2}

\item{data}{the data.frame where y, t, and x are}

\item{t}{conditional at a value T=t (i.e. should be scalar)}

\item{h}{bandwidth}
}
\value{
lige
}
\description{
local intergenerational elasticities
}
\keyword{internal}
