% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alige.R
\name{Plot_ALIGE}
\alias{Plot_ALIGE}
\title{Plot_ALIGE}
\usage{
Plot_ALIGE(ALIGE, sd_ALIGE)
}
\arguments{
\item{sd_ALIGE}{SD of ALIGE}
}
\value{
Plot of ALIGE
}
\description{
plot ALIGE with 95\% confidence intervals
}
