% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alige.R
\name{sdF}
\alias{sdF}
\title{sdF}
\usage{
sdF(B, tvals, Y, T, Xmat, h)
}
\arguments{
\item{B}{number of bootstrap iterations}

\item{tvals}{a grid of values of treatment variable}

\item{Y}{outcome variable}

\item{T}{treatment variable}

\item{Xmat}{covariates}

\item{h}{bandwidth}
}
\value{
sd
}
\description{
using wild bootstrap to obtain standard deviation
}
